rename: 페이지 경로와 파일 이름. 수정
