+++ src/components/sign-up/StepOne.tsx
@@ -1,0 +1,3 @@
+export const StepOne = () => {
+  return <div>StepOne</div>;
+};
