+++ src/pages/sign-up/individual/signup.module.scss
